table

+++ src/sample/core/ExponentialDistributionGenerator.java
@@ -2,12 +2,13 @@ package sample.core;
 
 import sample.system.Reference;
 
+import java.math.BigDecimal;
 import java.util.Random;
 
 public class ExponentialDistributionGenerator {
 
-    public static double generate() {
+    public static BigDecimal generate() {
         Random random = new Random();
-        return -(1 / Reference.LAMBDA_EXPONENTIAL) * Math.log(1 - random.nextDouble());
+        return BigDecimal.valueOf(-(1 / Reference.LAMBDA_EXPONENTIAL) * Math.log(1 - random.nextDouble()));
     }
 }
