+++ src/sample/core/NormaDistributionGenerator.java
@@ -2,12 +2,13 @@ package sample.core;
 
 import sample.system.Reference;
 
+import java.math.BigDecimal;
 import java.util.Random;
 
 public class NormaDistributionGenerator {
 
-    public static double generate() {
+    public static BigDecimal generate() {
         Random random = new Random();
-        return random.nextGaussian() * Reference.STANDARD_DEVIATION + Reference.MEAN;
+        return BigDecimal.valueOf(random.nextGaussian() * Reference.STANDARD_DEVIATION + Reference.MEAN);
     }
 }
