+++ src/sample/service/ServerService.java
@@ -5,17 +5,19 @@ import sample.core.ExponentialDistributionGenerator;
 import sample.core.NormaDistributionGenerator;
 import sample.domain.Issue;
 
+import java.math.BigDecimal;
+
 public class ServerService {
 
-    public static Double serve(Issue issue) {
+    public static BigDecimal serve(Issue issue) {
 
         if (issue.getType().equals(Issue.type.ERLANG)) {
-            double time = NormaDistributionGenerator.generate();
+            BigDecimal time = NormaDistributionGenerator.generate();
             issue.setAwaiting(false);
             return time;
 
         } else {
-            double time = ExponentialDistributionGenerator.generate();
+            BigDecimal time = ExponentialDistributionGenerator.generate();
             issue.setAwaiting(false);
             return time;
         }
